OTHER: Change unsigned to signed in output_t.
Some time ago we ditched support for unsigned
integers in the client library. Now output is
sanitized to only use gint.

The reason to use gint instead of gint32 is due
to g_atomic_* operations operate on gint, and
atomic operations could be used to remove some
unnecessary locking later.