FEATURE(1598): Extend genipc, empowering object methods.
Three attributes were added to method descriptions:

noreply: if "true", the server will waive the responsibility of
replying to the client that called the method, except when the method
returns an error.

need_client: if "true", the method will receive the calling client's id
in its parameters.

need_cookie: if "true", the method will receive the cookie associated
with the call as a parameter.